requirements.txt
smbus
python-mpd2
RPi.GPIO
Unidecode
